+++ scripts/appimage.sh
@@ -264,7 +264,7 @@ main() {
 
   # These could be symlinks but it seems they doesn't work with AppimageLauncher
   cp resources/icons/logo.svg Pragtical.AppDir/pragtical.svg
-  cp resources/linux/dev.pragtical.Pragtical.desktop Pragtical.AppDir/
+  cp resources/linux/dev.avi.Studio.desktop Pragtical.AppDir/
 
   if [[ $addons == true ]]; then
     addons_download "${build_dir}"